Fix incorrect JNI params in GA/GTI
The JNI Call(x)Method functions use va_args. GameActivity
was passing 64-bit int parameters to what should be
jint (32-bit signed integers), which was causing incorrect
parameter values to be passed via JNI on 32-bit ABIs.

Also adds a couple static_cast to jint in gametextinput for sign
correctness, and remove a spurious bool parameter from the
restartInputMethod JNI call.

Bug: b/296816862
Test: Modify AGDKTunnel sample as described in bug, build and test

+++ game-activity/prefab-src/modules/game-activity/include/game-activity/GameActivity.cpp
@@ -388,7 +388,8 @@ static int mainWorkCallback(int fd, int events, void *data) {
         case CMD_SET_WINDOW_FLAGS: {
             code->env->CallVoidMethod(code->javaGameActivity,
                                       gGameActivityClassInfo.setWindowFlags,
-                                      work.arg1, work.arg2);
+                                      static_cast<jint>(work.arg1),
+                                      static_cast<jint>(work.arg2));
             checkAndClearException(code->env, "setWindowFlags");
         } break;
         case CMD_SHOW_SOFT_INPUT: {
@@ -406,8 +407,10 @@ static int mainWorkCallback(int fd, int events, void *data) {
         case CMD_SET_IME_EDITOR_INFO: {
             code->env->CallVoidMethod(
                 code->javaGameActivity,
-                gGameActivityClassInfo.setImeEditorInfoFields, work.arg1,
-                work.arg2, work.arg3);
+                gGameActivityClassInfo.setImeEditorInfoFields,
+                static_cast<jint>(work.arg1),
+                static_cast<jint>(work.arg2),
+                static_cast<jint>(work.arg3));
             checkAndClearException(code->env, "setImeEditorInfo");
         } break;
         case CMD_RESTART_INPUT: {
